BASE-755 fix for all emtpy list delete repos

+++ src/main/kotlin/com/bitclave/node/repository/search/PostgresSearchRequestRepositoryImpl.kt
@@ -34,6 +34,7 @@ class PostgresSearchRequestRepositoryImpl(
     }
 
     override fun deleteByIdIn(ids: List<Long>): Int {
+        if (ids.isEmpty()) return 0
         repository.deleteTagsByIdIn(ids)
         return repository.deleteByIdIn(ids)
     }
